@Dieterbe 2021-11-02
hmm the benchmarks writes b.N points, which will often be artificially high.
also, the values aren't the most realistic..
a more real world benchmark would be useful.

are you seeing any benefit when deployed?