@shanson7 2021-11-02
> hmm the benchmarks writes b.N points, which will often be artificially high. 

Above benchmarks were run with `-benchtime=120x` to force 120 points per chunk

> also, the values aren't the most realistic.. a more real world benchmark would be useful.
> are you seeing any benefit when deployed?

We ran profiles before and after and saw the expected drop in time spent in `Push`. However, we didn't really see any change in blackbox metrics (CPU usage, ingest rate) likely because such a small percentage of time was spent in `Push`.