@Dieterbe 2021-11-02
> Above benchmarks were run with -benchtime=120x to force 120 points per chunk

with similar results?