@petethepig 2021-11-18
FWIW we ran one of the benchmarks (`PushSeriesLong`) with pyroscope (I'm one of the maintainers) and here's a diff flamegraph:

[![Pasted_Image_11_17_21__5_17_PM](https://user-images.githubusercontent.com/662636/142335557-0f8a0844-6669-4443-8d03-3c5e4ad5140f.png)](https://metrictank.pyroscope.cloud/comparison-diff?from=1637195510&leftFrom=1637195530&leftUntil=1637195620&rightFrom=1637195760&rightUntil=1637195830&until=1637195860&query=metrictank.cpu%7B%7D)

Green is functions that are faster and red is functions that are slower after the PR. We ran this with 8,000,000 iterations (to get enough profiling samples) inserting 120 data points at a time. 

You can see from the diff that calls to `writeBit` and `writeByte` happen much less often now, and more work is done in `writeBites` (e.g the new `alignByte` call). Overall you can see about 30% improvement for the whole benchmark (goes from 1.20 mins to 0.83 mins).

Here's a [link to a publicly available pyroscope instance](https://metrictank.pyroscope.cloud/comparison-diff?from=1637195510&leftFrom=1637195530&leftUntil=1637195620&rightFrom=1637195760&rightUntil=1637195830&until=1637195860&query=metrictank.cpu%7B%7D) with this data) if you want to play with this profiling data.

I do think running pyroscope on some production cluster would be more interesting and produce more representative results. @shanson7, @Dieterbe — I would love to collaborate on that if any of you are interested.