@Dieterbe 2021-11-27
So:
1) In `writeByte()` you tweaked the optimal path: when the stream ends at at a byte boundary, we can skip some instructions and don't append an anticipatory byte at the end. This might be helpful when the last write to the stream is such an "optimal path write".
2) In master, in `writeBits()` we write whole bytes as long as we have any, and then all remaining bits, individually (ignoring byte alignment). Your change does a (likely) partial byte write to achieve alignment, then writes whole bytes, and finally the remaining bits in one shot.

The code looks fine, but I have yet to check out #2005 (I see that the benchmarks quoted here are introduced there).  Are your benchmarks against #2005 or against master? There's a couple distinct optimizations in this PR, and the workload will change due to #2005, so let me check out that PR to make more sense of this. :)

From only looking at the code, I'm pretty confident that 1) will always be a net improvement, though i'm doubtful by how much, 2) is likely also an improvement, though it's less obvious.