@shanson7 2021-11-29
> Are your benchmarks against #2005 or against master?

Hmm, that's a good question. It's been a while and I cannot remember if I ported the benchmarks or if I based this whole changed on #2005.